v0.0.40-alpha ⬇︎
the exporter.js it is an helper file to export properties and propertiesSymbol.
due to the nature of Symbol and the way how I instatiate them (with description text of the property being added to the main Object)
it's necessary to export them first and then in the index.js exporting the <value:function>|<value:data> of the property.